+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Adds oihana\files\countFileLines
 - Adds oihana\files\getFileLines
 - Adds oihana\files\getFileLinesGenerator
+- Adds the .cbor 'application/cbor' file mime type and file extension constants.
 
 ## [1.0.0] - 2025-08-13